adapter/server: renamed listener adapter to server adapter
