+++ adapter/server/http_test.go
@@ -1,4 +1,4 @@
-package listener_test
+package server_test
 
 import (
 	. "github.com/onsi/ginkgo"
